+++ packages/altive_lints_plugin/lib/altive_lints_plugin.dart
@@ -1,0 +1,4 @@
+/// Analyzer Plugin implementation used by `package:altive_lints`.
+library;
+
+export 'main.dart' show plugin;